+++ src/AppBundle/Controller/API/ProjectCategoryController.php
@@ -4,6 +4,7 @@ namespace AppBundle\Controller\API;
 
 use AppBundle\Entity\ProjectCategory;
 use AppBundle\Form\ProjectCategory\CreateType;
+use AppBundle\Security\ProjectVoter;
 use MainBundle\Controller\API\ApiController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -78,6 +79,8 @@ class ProjectCategoryController extends ApiController
      */
     public function getAction(ProjectCategory $projectCategory)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::VIEW, $projectCategory->getProject());
+
         return $this->createApiResponse($projectCategory);
     }
 
@@ -94,6 +97,8 @@ class ProjectCategoryController extends ApiController
      */
     public function editAction(Request $request, ProjectCategory $projectCategory)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::EDIT, $projectCategory->getProject());
+
         $data = $request->request->all();
         $form = $this->createForm(CreateType::class, $projectCategory, ['csrf_protection' => false]);
         $form->submit($data, false);
@@ -128,6 +133,8 @@ class ProjectCategoryController extends ApiController
      */
     public function deleteAction(ProjectCategory $projectCategory)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::DELETE, $projectCategory->getProject());
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($projectCategory);
         $em->flush();
